koppla post till databasen och lägg till user-post relationer

+++ src/main/java/se/jensen/sofi_n/social_app/dto/PostResponseDTO.java
@@ -2,4 +2,4 @@ package se.jensen.sofi_n.social_app.dto;
 
 import java.time.LocalDateTime;
 
-public record PostResponseDTO(int id, String text, LocalDateTime createdAt) {}
+public record PostResponseDTO(Long id, String text, LocalDateTime createdAt) {}
